requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
sumy==0.11.0
nltk==3.8.1
python-dotenv==1.0.0
newspaper3k==0.2.8 